pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "jobfinder"
version = "0.1.0"
description = "LinkedIn job search scraper and SQLite pipeline MVP"
requires-python = ">=3.9"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "requests>=2.32.3",
  "PyYAML>=6.0.2",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["jobfinder*"]
